Language log fix (#290)
* Fix log file continuously duplicating error messages(#288)

Change it to record only once.
If your native language is English, you may not find this problem. I used an external translation file json to translate the menu into my language. When I open the menu, an error message will be written to the menyooLog file for each frame. As long as you stay on the current menu page, it will repeatedly write the same message for each menu line, like this:

* moving from filelogger.cpp to language.cpp instead

Restore filelogger.cpp

* UpVersion for Pre-release

---------

Co-authored-by: lucienlmy <101241027+lucienlmy@users.noreply.github.com>