New Blip features (#292)
* Update GTAblip.cpp

⚠VERY EXPERIMENTAL⚠

I have tried to add some new features to blips. Please recognize that I have never coded in my life and this is all black magic to me.

It would mean a lot to me if this could be tested!

Added;
Toggle for cones (like NPC vision, likely won't function but it's there for the aesthetic).

Toggle to make blips not selectable on the map.

Set blip priority.

Toggle to allow blip rotation.

Toggle (hopefully just a toggle) to have a blip to mark an area.

Changes;
Moved Enemy, Friend, and VIP to the top

* Update GTAblip.h

⚠VERY EXPERIMENTAL⚠
Additions made to this code were added by someone with little to no coding experience.

See commit to GTAblip.cpp for a full list of changes.

* Update GTAblip.h

* Added one last component.

Added the function for blips to rotate in sync with the attached entity (in theory).

* Update STSTasks.cpp

Adding handles.

* Update STSTasks.h

* Update Submenus_TaskSequence.cpp