Update GTAentity.cpp
While I try to figure out Github Desktop and attempt to fix the conflicts in my Bodyguard script I added a small line to include vehicle engine health.

In GTA 5, vehicles have two health bars, the first barely does anything, the second is the engine health. With this we can now control the health of a vehicle's engine.